// src/app/receipts/models/creditNote/create-creditNote.model.ts
import { GetSelectListModel } from "@shared/models/get-select-list.model";
import { GetBankAccountModel } from "src/app/master-data/models/bank/get-bank-account.model";
import { GetCustomerModel } from "src/app/master-data/models/customer/get-customer-model";
import { GetCashBoxModel } from '../cashbox/get-cash-box.model';
import { CreatePaymentTransactionModel } from "../paymentTransaction/create-payment-transaction.model";
import { CreateCostElementItemModel } from "../costelement/create-cost-element-item.model";

export interface CreateCreditNoteModel {
  id: number;
  code: string;
  voucherTypeId: string;
  documentDate: string;
  refNumber: string;
  paymentOnlineRef: string;
  isPaidOnline: boolean;
  customerId: number;
  entityCode: string;
  salesRepresentativeId: string;
  arabicRemarks: string;
  englishRemarks: string;
  creditCardTypeId: string;
  discountCouponId: string;
  isBankDeposit: boolean;
  bankAccountId: string;
  qrCode: string;
  bankDepositAmount: number;
  isCashBox: boolean;
  isArchived: boolean;
  cashBoxId: string;
  cashBoxAmount: number;
  netValueAfterDiscount: number;
  taxAmount: number | null;
  netValueAfterTax: number | null;
  netValue: number;
  tolalPaid: number;
  totalRefund: number;
  sectorTypeId: string;
  branchId: string;
  isDownPayment: boolean;
  creditCardType: GetSelectListModel;
  customer: GetCustomerModel;
  bankAccount: GetBankAccountModel;
  cashBox: GetCashBoxModel;
  paymentsTransactions: CreatePaymentTransactionModel[];
  costElements: CreateCostElementItemModel[];
  totalTaxAmount: number;
}